Move next_announcement_after to model
It is basically a constructor to `BitMexPriceEventId` which is already
part of `model::olivia`.
