+++ daemon/src/oracle.rs
@@ -6,16 +6,14 @@ use async_trait::async_trait;
 use futures::StreamExt;
 use maia_core::secp256k1_zkp::schnorrsig;
 use model::olivia;
+use model::olivia::next_announcement_after;
 use model::olivia::BitMexPriceEventId;
 use model::CfdEvent;
 use model::EventKind;
 use std::collections::HashMap;
 use std::collections::HashSet;
-use std::ops::Add;
-use time::ext::NumericalDuration;
 use time::Duration;
 use time::OffsetDateTime;
-use time::Time;
 use tokio_tasks::Tasks;
 use xtra_productivity::xtra_productivity;
 use xtras::SendInterval;
@@ -309,20 +307,6 @@ impl Actor {
 #[error("Announcement {0} not found")]
 pub struct NoAnnouncement(pub BitMexPriceEventId);
 
-pub fn next_announcement_after(timestamp: OffsetDateTime) -> BitMexPriceEventId {
-    let adjusted = ceil_to_next_hour(timestamp);
-
-    BitMexPriceEventId::with_20_digits(adjusted)
-}
-
-fn ceil_to_next_hour(original: OffsetDateTime) -> OffsetDateTime {
-    let timestamp = original.add(1.hours());
-    let exact_hour = Time::from_hms(timestamp.hour(), 0, 0).expect(
-        "Exact hour from timestamp to be always within range, both docs and tests confirm it",
-    );
-    timestamp.replace_time(exact_hour)
-}
-
 #[async_trait]
 impl xtra::Actor for Actor {
     type Stop = ();
@@ -388,30 +372,4 @@ impl Attestation {
 
 impl xtra::Message for Attestation {
     type Result = ();
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use time::macros::datetime;
-
-    #[test]
-    fn next_event_id_after_timestamp() {
-        let event_id = next_announcement_after(datetime!(2021-09-23 10:40:00).assume_utc());
-
-        assert_eq!(
-            event_id.to_string(),
-            "/x/BitMEX/BXBT/2021-09-23T11:00:00.price?n=20"
-        );
-    }
-
-    #[test]
-    fn next_event_id_is_midnight_next_day() {
-        let event_id = next_announcement_after(datetime!(2021-09-23 23:40:00).assume_utc());
-
-        assert_eq!(
-            event_id.to_string(),
-            "/x/BitMEX/BXBT/2021-09-24T00:00:00.price?n=20"
-        );
-    }
 }
